inforce stricker limits on age
Prior to this change there was no lower bound or upper bound on age.
That would allow negative ages and ages greater than 216 months (18
years) to give non-zero cardiovascular and non-zero renal scores.

This change sets the logic in R, python, and SQL, to use age > 0 and age
<= 216.   Tables in the vignette and man files have been updated as
well.

fix #3