Update all references to the data paper with a note for equal contributions between Sanchez-Pinto and Bennet; DeWitt and Russell.
Prior to this change all the references where just copied from the "Cite
This" provided on the JAMA website.

This change add Russell explicitly and asstricks and a note about the
contributions.  The language is copied form the JAMA paper as well.

There is one exception - the DESCRIPTION file does not have these
indicators for contribution levels.

re #7