Improving documentation of reference materials
Refactor how the two JAMA papers are cited - all the authors and equal
contribution notes are in the phoenix-package documentation with links
and directions to review the details there in all the other function
manual files.   The CITATION file as been extended to include notes on
the two JAMA papers and instructions on how to cite the R package.

re #7, #8